+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/RequestParametersFrom.kt
@@ -20,13 +20,16 @@ sealed class RequestParametersFrom<S : RequestParameters> {
         override val parameters: T,
         @SerialName(SerialNames.VERIFIED)
         val verified: Boolean,
+        @SerialName(SerialNames.PARENT)
+        val parent: Url?,
     ) : RequestParametersFrom<T>() {
         override fun toString(): String {
-            return "JwsSigned(jwsSigned=${jwsSigned.serialize()}, parameters=$parameters, verified=$verified)"
+            return "JwsSigned(parent='$parent', jwsSigned=${jwsSigned.serialize()}, parameters=$parameters, verified=$verified)"
         }
     }
 
     @Serializable
+    @SerialName(SerialNames.TYPE_DCAPI_SIGNED)
     data class DcApiSigned<T : RequestParameters>(
         @SerialName(SerialNames.DC_API_REQUEST)
         val dcApiRequest: Oid4vpDCAPIRequest,
@@ -42,6 +45,7 @@ sealed class RequestParametersFrom<S : RequestParameters> {
     }
 
     @Serializable
+    @SerialName(SerialNames.TYPE_DCAPI_UNSIGNED)
     data class DcApiUnsigned<T : RequestParameters>(
         @SerialName(SerialNames.DC_API_REQUEST)
         val dcApiRequest: Oid4vpDCAPIRequest,
@@ -75,20 +79,25 @@ sealed class RequestParametersFrom<S : RequestParameters> {
         val jsonString: String,
         @SerialName(SerialNames.PARAMETERS)
         override val parameters: T,
+        @SerialName(SerialNames.PARENT)
+        val parent: Url?,
     ) : RequestParametersFrom<T>() {
         override fun toString(): String {
-            return "Json(jsonString='$jsonString', parameters=$parameters)"
+            return "Json(parent='$parent', jsonString='$jsonString', parameters=$parameters)"
         }
     }
 
     object SerialNames {
         const val TYPE_JWS_SIGNED = "JwsSigned"
         const val TYPE_JSON = "Json"
+        const val TYPE_DCAPI_UNSIGNED = "DcApiUnsigned"
+        const val TYPE_DCAPI_SIGNED = "DcApiSigned"
         const val TYPE_URI = "Uri"
 
         const val JWS_SIGNED = "jwsSigned"
         const val JSON_STRING = "jsonString"
         const val URL = "url"
+        const val PARENT = "parent"
         const val PARAMETERS = "parameters"
         const val DC_API_REQUEST = "dcApiRequest"
         const val VERIFIED = "verified"