+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/SimpleAuthorizationService.kt
@@ -292,7 +292,7 @@ class SimpleAuthorizationService(
     private suspend fun RequestParameters.extractPushedRequestParams() = when (this) {
         is JarRequestParameters -> {
             require(requestUri == null) { "request_uri must not be set for PAR" }
-            requestParser.extractRequest(this)?.parameters as? AuthenticationRequestParameters
+            requestParser.extractRequest(this, null)?.parameters as? AuthenticationRequestParameters
                 ?: throw InvalidRequest("request must contain valid authorization request parameters")
         }
 
@@ -368,7 +368,7 @@ class SimpleAuthorizationService(
                 if (clientId != input.clientId)
                     throw InvalidRequest("client_id not matching from par: ${input.clientId} vs $clientId")
             }
-        } ?: (requestParser.extractRequest(input)?.parameters as? AuthenticationRequestParameters)
+        } ?: (requestParser.extractRequest(input, null)?.parameters as? AuthenticationRequestParameters)
         ?: throw InvalidRequest("could not parse request object from request")
 
         is RequestObjectParameters -> throw InvalidRequest("could not parse request object from request")